requirements.txt
vtracer==0.6.11
rembg==2.0.59
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.9.0.80